rename GlobalGrid related files
